pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "os-app"
version = "0.0.1"
requires-python = ">=3.8"
dependencies = [
    "Flask==3.0.2",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Flask-WTF==1.2.1",
    "Flask-Migrate==4.0.5",
    "Flask-Mail==0.9.1",
    "email-validator==2.1.0.post1",
    "python-dotenv==1.0.1",
    "Werkzeug==3.0.1",
    "qrcode[pil]==7.4.2",
    "WeasyPrint==62.1",
    "cryptography==42.0.8",
    "pyOpenSSL==24.1.0",
    "pytest"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "bandit==1.7.5",
    "safety==2.3.5",
    "pre-commit==3.5.0"
]

[tool.setuptools]
py-modules = ["app", "wsgi"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
# Exclude conftest.py from Black formatting - we handle it specially
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
  | instance
  | db
  # files
  | tests/conftest\.py
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["app", "models", "routes", "utils"]
known_third_party = ["flask", "sqlalchemy", "pytest"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["tests/conftest.py"]
